Create model to represent coordinates

+++ TruckPad/app/src/main/java/com/truckpad/androidcase/model/RouteRequest.kt
@@ -3,7 +3,7 @@ package com.truckpad.androidcase.model
 import com.google.gson.annotations.SerializedName
 
 data class RouteRequest(
-    @SerializedName("fuel_consumption") val fuelConsumption: Int,
+    @SerializedName("fuel_consumption") val fuelConsumption: Double,
     @SerializedName("fuel_price") val fuelPrice: Double,
     @SerializedName("places") val places: List<Place>
 )
